requirements.txt
argparse
PyYAML
requests
bs4